Core: generalise capture frames to f32 with a real/complex layout
Phase 10.0 D1b: ChannelCapture.data is f32 with per-component IqCal, and
CaptureFrame carries the frame-level SampleLayout. Scope backends convert
their i8 +/-125 wire encoding at frame construction, so scope behaviour is
unchanged. .nwc is v3 (layout + IqCal; v1/v2 still read). The GPU buffer is
array<f32> with NaN gap sentinels and NO_DATA is f32::NAN.

Sim only. All tests, WGSL validation, and the windowed pixel, geometry and
flow tests pass. Budget 33 files / net 362 lines (cap recorded at 35).

+++ crates/neowon-dsp/src/decode/i2c.rs
@@ -105,12 +105,12 @@ mod tests {
     use super::*;
 
     /// Build SCL/SDA waveforms for a write transaction.
-    fn transaction(addr: u8, data: &[u8], per_bit: usize) -> (Vec<i8>, Vec<i8>) {
+    fn transaction(addr: u8, data: &[u8], per_bit: usize) -> (Vec<f32>, Vec<f32>) {
         let (mut scl, mut sda) = (Vec::new(), Vec::new());
         let mut push = |c: bool, d: bool, n: usize| {
             for _ in 0..n {
-                scl.push(if c { 100i8 } else { -100 });
-                sda.push(if d { 100i8 } else { -100 });
+                scl.push(if c { 100.0 } else { -100.0 });
+                sda.push(if d { 100.0 } else { -100.0 });
             }
         };
         // Idle high, then START: SDA falls while SCL is high.
